+++ sk-merge/internal/config/config.go
@@ -25,7 +25,7 @@ type ClientProviderConfig struct {
 	GroupAuthority      *bool         `yaml:"groupAuthority"`      // Group will be fetched. Default true
 	Critical            *bool         `yaml:"critical"`            // If true (default), a failure on this provider will leads 'invalid login'. Even if another provider grants access
 	GroupPattern        string        `yaml:"groupPattern"`        // Group pattern. Default "%s"
-	UidOffset           int64         `yaml:"uidOffset"`           // Will be added to the returned Uid. Default to 0
+	UidOffset           int           `yaml:"uidOffset"`           // Will be added to the returned Uid. Default to 0
 }
 
 type ServiceConfig struct {
